fix(tests-linux): SHUT_RDWR is Int on Glibc — cast to Int32 for shutdown()
The Linux build now compiles past the socket-type fixes but failed at
`LoopbackHarness.swift:245`: `shutdown()`'s `how` argument is `Int32`, and
`SHUT_RDWR` imports as a plain `Int` on Glibc (it is `Int32` on Darwin).
`Int32(SHUT_RDWR)` is correct on both. A full grep of the harness confirms this
was the last un-normalized socket constant/syscall in the file.

+++ Tests/ADServeCoreTests/LoopbackHarness.swift
@@ -242,7 +242,9 @@ final class TestSocket: @unchecked Sendable {
     }
 
     func close() {
-        _ = shutdown(descriptor, SHUT_RDWR)
+        // `SHUT_RDWR` is a plain `Int32` on Darwin but an `Int` on Glibc — normalize to the `Int32`
+        // `shutdown()`'s `how` argument wants.
+        _ = shutdown(descriptor, Int32(SHUT_RDWR))
     }
 }
